# Optimize Backward Time Complexity to O(MK)