--- a/src/caffe/layers/center_loss_layer.cu
+++ b/src/caffe/layers/center_loss_layer.cu
@@ -7,8 +7,11 @@
 namespace caffe {
 
 template <typename Dtype>
-__global__ void Compute_distance_data_gpu(int nthreads, const int K, const Dtype* bottom,
-	      const Dtype* label, const Dtype* center, Dtype* distance) {
+__global__ void Compute_distance_data_gpu(int nthreads, const int K,
+                                          const Dtype* bottom,
+                                          const Dtype* label,
+                                          const Dtype* center,
+                                          Dtype* distance) {
   CUDA_KERNEL_LOOP(index, nthreads) {
     int m = index / K;
     int k = index % K;
@@ -17,55 +20,69 @@ __global__ void Compute_distance_data_gpu(int nthreads, const int K, const Dtype
     distance[index] = bottom[index] - center[label_value * K + k];
   }
 }
-
 template <typename Dtype>
-__global__ void Compute_center_diff_gpu(int nthreads, const int M, const int K, 
-        const Dtype* label, const Dtype* distance, Dtype* variation_sum, 
-        Dtype* center_diff) {
+__global__ void Compute_variation_sum_gpu(int nthreads, const int K,
+                                          const Dtype* label,
+                                          const Dtype* distance,
+                                          Dtype* variation_sum, int * count) {
   CUDA_KERNEL_LOOP(index, nthreads) {
-    int count = 0;
-    for (int m = 0; m < M; m++) {
-      const int label_value = static_cast<int>(label[m]);
-      if (label_value == index) {
-        count++;
-        for (int k = 0; k < K; k++) {
-          variation_sum[index * K + k] -= distance[m * K + k];
-        }
-      }
-    }
-    for (int k = 0; k < K; k++) {
-      center_diff[index * K + k] = variation_sum[index * K + k] /(count + (Dtype)1.);
-    }
+    int m = index / K;
+    int k = index % K;
+    const int label_value = static_cast<int>(label[m]);
+    variation_sum[label_value * K + k] -= distance[m * K + k];
+    count[label_value] +=  ((k == 0)?1:0);
   }
 }
-
-
+  template <typename Dtype>
+__global__ void Compute_center_diff_gpu(int nthreads, const int K,
+                                        const Dtype* label,
+                                        Dtype* variation_sum,
+                                        int * count, Dtype* center_diff) {
+    CUDA_KERNEL_LOOP(index, nthreads) {
+      int m = index / K;
+      int k = index % K;
+      const int n = static_cast<int>(label[m]);
+      center_diff[n * K + k] = variation_sum[n * K + k]
+        / (count[n] + (Dtype)1.);
+    }
+}
 template <typename Dtype>
 void CenterLossLayer<Dtype>::Forward_gpu(const vector<Blob<Dtype>*>& bottom,
     const vector<Blob<Dtype>*>& top) {
   int nthreads = M_ * K_;
-  Compute_distance_data_gpu<Dtype><<<CAFFE_GET_BLOCKS(nthreads),
-      CAFFE_CUDA_NUM_THREADS>>>(nthreads, K_, bottom[0]->gpu_data(), bottom[1]->gpu_data(),
-                                this->blobs_[0]->gpu_data(), distance_.mutable_gpu_data());
+  Compute_distance_data_gpu<Dtype> <<< CAFFE_GET_BLOCKS(nthreads),
+      CAFFE_CUDA_NUM_THREADS>>>(nthreads, K_, bottom[0]->gpu_data(),
+                                bottom[1]->gpu_data(),
+                                this->blobs_[0]->gpu_data(),
+                                distance_.mutable_gpu_data());
   Dtype dot;
   caffe_gpu_dot(M_ * K_, distance_.gpu_data(), distance_.gpu_data(), &dot);
   Dtype loss = dot / M_ / Dtype(2);
   top[0]->mutable_cpu_data()[0] = loss;
 }
-
 template <typename Dtype>
 void CenterLossLayer<Dtype>::Backward_gpu(const vector<Blob<Dtype>*>& top,
     const vector<bool>& propagate_down,
     const vector<Blob<Dtype>*>& bottom) {
-  int nthreads = N_;
-  caffe_gpu_set(N_ * K_, (Dtype)0., variation_sum_.mutable_cpu_data());
-  Compute_center_diff_gpu<Dtype><<<CAFFE_GET_BLOCKS(nthreads),
-      CAFFE_CUDA_NUM_THREADS>>>(nthreads, M_, K_, bottom[1]->gpu_data(), distance_.gpu_data(), 
-                                variation_sum_.mutable_cpu_data(), this->blobs_[0]->mutable_gpu_diff());
-
+  caffe_gpu_set(N_ * K_, (Dtype)0., variation_sum_.mutable_gpu_data());
+  caffe_gpu_set(N_, 0 , count_.mutable_gpu_data());
+  caffe_gpu_set(N_ * K_, (Dtype)0., this->blobs_[0]->mutable_gpu_diff());
+  int nthreads = M_ * K_;
+  Compute_variation_sum_gpu<Dtype> <<< CAFFE_GET_BLOCKS(nthreads),
+    CAFFE_CUDA_NUM_THREADS>>>(nthreads, K_, bottom[1]->gpu_data(),
+                              distance_.gpu_data(),
+                              variation_sum_.mutable_gpu_data(),
+                              count_.mutable_gpu_data());
+  Compute_center_diff_gpu<Dtype> <<< CAFFE_GET_BLOCKS(nthreads),
+    CAFFE_CUDA_NUM_THREADS>>>(nthreads, K_, bottom[1]->gpu_data(),
+                              variation_sum_.mutable_gpu_data(),
+                              count_.mutable_gpu_data(),
+                              this->blobs_[0]->mutable_gpu_diff());
   if (propagate_down[0]) {
-    caffe_gpu_scale(M_ * K_, top[0]->cpu_diff()[0] / M_, 
-                             distance_.gpu_data(), bottom[0]->mutable_gpu_diff());
+    caffe_gpu_scale(M_ * K_,
+                    top[0]->cpu_diff()[0] / M_,
+                    distance_.gpu_data(),
+                    bottom[0]->mutable_gpu_diff());
   }
   if (propagate_down[1]) {
     LOG(FATAL) << this->type()