@mfs6174 2016-10-24
Hi, @jiangxuehan 

I am still working on that with both paper author's code and my code.

I have only tested my PR's code with the MNIST toy example. Starting from the same snapshot and training data (no shuffle during training), my code produced exactly the same center_diff values and nearly the same test result with the author's code.

Which code did you trained the model with?  Was it the paper author's code or my PR's code? If you can reproduce the result with the paper author's code but not my code, I will check my code again. If you can neither reproduce the result with the paper author's code, I will discuss with you  further when I finish my experiments for reproducing the result on LFW.