@chichan01 2016-10-24
Hi  @jiangxuehan and @mfs6174,
  I use MTCNN to get the 5pts landmarks and use his provided model. The accuracy  in LFW in view 2 is just 96.55 +/-0.229129.  If I use his provided feature, the accuracy is 98.98 +/- 0.186685.
If I use other way to get 5pts landmarks and use his provide model, result is 98.75 +/-0.194754.
If I train from scratch using my 5 pts landmarks, the result is 98.47 +/- 0.211986.
  I hope this will give you some guidelines.  Besides, I found that their system cannot use dropout as the dropout will make the loss to NAN. Do you guys have the same problem when dropout is applied?