@jiangxuehan 2016-10-24
Hi, @mfs6174  @chichan01 ,
to mfs6174 :  
I used the paper author's code. Further discussions after you finish experiments  will be appreciated. (By the way,your optimized code is correct and faster)
to chichan01:  
Your result seems reasonable. Did you use the same code\model\data as the author's PR? Have you ever made some changes except landmarks?
I have tried to use dropout for the 512d-fc layer, the loss didn't produce NAN value. I think center loss acts as a regular term , it's more reasonable to compare softmax+dropout vs. softmax+center_loss. Have you tried softmax+dropout?