@chichan01 2016-10-24
Hi @jiangxuehan,
   I only try to reproduce their result in this moment and therefore I did not change anything.  I did try to apply centre loss in other architecture but it seems that it need to tune loss_weight. Anyway, I will try to compare it with dropout.