@jiangxuehan 2016-10-24
@chichan01 : 
Could you please send your train log files to me(both 98.47 and NAN  )? I want to compare loss curves and hope it can provide some useful information. My email is nicehjs@163.com, Thanks.