@chichan01 2016-10-25
Hi @jiangxuehan and @mfs6174,
  I directly use  author's prototxt as their network in prototxt is not similar to their paper, therefore, I think following the description in the paper may not right.  Certainly, if you test it based on their description, please let me know the result. Also, I would like to thank for telling me that how I get the alpha.
  To @jiangxuehan ,  I am sorry that I am not able to give you the log file as the log files were corrupted.
By the way, did you guys able to reproduce their megaface result?