@jiangxuehan 2016-10-25
Hi @mfs6174  @chichan01 
As mfs6174 metioned, center = center - lr_mult*lr*d(center), so alpha = lr_mult*lr.With lr decrease from 0.1 to 0.01/0.001, we should set lr_mult to 5/50/500 theoretically. Besides , weight_decay should be set to 0. Another method is rewrite the Backward function : center = center - alpha*d(center).  If @chichan01 can reproduce the result with this PR, maybe alpha is not so important?
(I have not do any experiment about alpha, above content is just my own comprehension. Could @ydwen explain about alpha for us? Thanks.)

To chichan01: I am still working on LFW, and meeting some obsessions to get ~99% accuracy.  How about your experiment on Megaface?