@mfs6174 2016-10-25
Hi,

I am still working on that with both paper author's code and my code.

I have only tested my PR's code with the MNIST toy example. Starting from
the same snapshot and training data (no shuffle during training), my code
produced exactly the same center_diff values and nearly the same test
result with the author's code.

Which code did you trained the model with?  Was it the paper author's code
or my PR's code? If you can reproduce the result with the paper author's
code but not my code, I will check my code again. If you can neither
reproduce the result with the paper author's code, I will discuss with you
 further when I finish my experiments for reproducing the result on LFW.

2016-10-24 8:59 GMT+08:00 jiangxuehan <notifications@github.com>:

> @mfs6174 <https://github.com/mfs6174> , have you reproduced the result on
> LFW or Megaface? I have trained the model on CASIA and tested on LFW ,but
> it didn't work well (EER~96.5%).
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/ydwen/caffe-face/pull/1#issuecomment-255627529>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAvOsgRKov1iq31hSBSj-KV03zDjR8tvks5q3AL8gaJpZM4KdJSj>
> .
>



-- 
--------------------
mfs6174
Web  mfs6174.org