@ydwen 2016-10-27
@jiangxuehan @chichan01 @ @mfs6174 
Sorry for replying this late. I was quite busy in these days. Thanks for the commit from @mfs6174, I will check it ASAP.

Here I try to answer some of your questions. If anything is not clear, please feel free to let me know.

Common issues
Dropout:  if you want to combine dropout, the current code may not support. Concretely, if some of the elements in x is omitted, the corresponding elements in center should be omit as well. If not, the loss becomes unstable.

Alpha: At the beginning, we use the update strategy as in paper. It performs well in our experiments (deepidnet & mixed dataset). When refactoring the code, we find that the implementation of updating alpha in caffe is not elegant and becomes complicated with multi-gpu. Finally we try lr and weight_decay in caffe and find it is very convenient and works pretty well (achieving near the same performance as before). So you guys can try modifying the lr and weight_decay for different alphas.

Network Architecture: At the beginning we use the deepidnet so we have to implement the local convolution layer. We are not going to released our implementation since it is quite inelegant (^_^). The most important is, local convolution layer is complicated and not efficient in time (~8h for resnet with 28 layers v.s. ~14h for deepidnet with 6 layers) and space (~100M for resnet with 28 layers v.s. ~200M for deepidnet with 6 layers). Therefore, we use a modified resnet as our network. It performs better with less parameters and training time.

To @chichan01 , I guess the patch you used is not the same as in the demo or the given model. Please double check the positions of eyes, nose, mouth corner in the cropped face and use our provided template (5 points landmarks) given in demo.

To @chichan01 , loss weight is related to the total number of classes, i.e. num_output in fc6. Generally speaking, the more classes, the smaller loss weight is.

To @jiangxuehan Please provide more details of your experiment (96.5% EER), or I can't give you some hints.